+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Usprawnienia do `rethrow_as`
 * Travis teraz wrzuca na PyPI, a nie ja :)
 * Dodano `typednamedtuple`
+* Poprawione zachowanie `coerce`
 
 ## v2.0.14